+++ src/main/groovy/ru/vyarus/gradle/plugin/pom/PomExtension.groovy
@@ -1,9 +1,10 @@
 package ru.vyarus.gradle.plugin.pom
 
 import groovy.transform.CompileStatic
-import groovy.transform.TypeCheckingMode
-import org.gradle.api.GradleException
+import org.gradle.api.Action
 import org.gradle.api.Project
+import org.gradle.api.XmlProvider
+import org.gradle.api.publish.maven.MavenPom
 
 /**
  * Pom modifications configuration.
@@ -11,18 +12,40 @@ import org.gradle.api.Project
  * Example usage:
  * <pre>
  * <code>
- *     maven.pom {
- *         licenses {
- *              license {
- *                  name "The MIT License"
- *              }
- *         }
- *         developers {
- *             developer {
- *                 id "dev1"
- *                 name "Dev1 Name"
- *                 email "dev1@email.com"
+ *     maven {
+ *        // static model - same as pom in publication (exactly the same!)
+ *        pom {
+ *             licenses {
+ *                  license {
+ *                      name = "The MIT License"
+ *                  }
  *             }
+ *             developers {
+ *                 developer {
+ *                     id = "dev1"
+ *                     name = "Dev1 Name"
+ *                     email = "dev1@email.com"
+ *                 }
+ *             }
+ *         }
+ *         // raw closure (no structure limitations) - old configuration way
+ *         withPom {
+ *             licenses {
+ *                  license {
+ *                      name "The MIT License"
+ *                  }
+ *             }
+ *             developers {
+ *                 developer {
+ *                     id "dev1"
+ *                     name "Dev1 Name"
+ *                     email "dev1@email.com"
+ *                 }
+ *             }
+ *         }
+ *         // exactly the same as publication's pom.withXml
+ *         withPomXml {
+ *             asNode().appendNode('tata', 'blabla')
  *         }
  *     }
  * </code>
@@ -49,12 +72,13 @@ class PomExtension {
     boolean disabledScopesCorrection
     boolean disabledBomsReorder
 
-    protected List<Closure> configs = []
-    protected List<Closure> xmlModifiers = []
+    protected List<Action<? super MavenPom>> configs = []
+    protected List<Closure> rawConfigs = []
+    protected List<Action<? super XmlProvider>> xmlModifiers = []
 
     // in multi-module project used for incorrect initialization detection when sub module's convention
     // wasn't initialized and root module's convention used (hard to track problems)
-    private String projectPath
+    private final String projectPath
 
     PomExtension(Project project) {
         // project name where convention was registered
@@ -109,54 +133,35 @@ class PomExtension {
     }
 
     /**
-     * @param config user pom
+     * Exactly the same as pom configuration in publication, but applied to all publications.
+     * <p>
+     * ATTENTION: All properties assignment must be done with '='
+     *
+     * @param config configuration action
      */
-    void pom(Closure config) {
-        validateTargetProjectCorrectness('pom', config)
+    void pom(Action<? super MavenPom> config) {
         this.configs.add(config)
     }
 
     /**
-     * Modification closure is called just after user pom merge. Pom xml passed to closure as {@link Node} parameter.
-     * @param modifier manual pom modification closure
+     * Pure groovy closure configuration (old way). Does not limited with any structure and so ANY xml tags could
+     * be specified like this.
+     * <p>
+     * ATTENTION: All properties assignment must be done as "method" - does not contain '='!
+     *
+     * @param config user pom
      */
-    void withPomXml(Closure modifier) {
-        validateTargetProjectCorrectness('withPomXml', modifier)
-        this.xmlModifiers.add(modifier)
-    }
-
-    private void validateTargetProjectCorrectness(String type, Closure config) {
-        String srcProject = findDeclarationProjectPath(config)
-        // catching mistakes of root project configuration from module, but allowing subproject configuration
-        // from upper project (last condition): project(':mod').pom { ... }
-        if (srcProject && projectPath != srcProject && !projectPath.startsWith(srcProject)) {
-            throw new GradleException("${type.capitalize()} closure declared in project '$srcProject'" +
-                    " actually applied in '$projectPath' project (conventions are searched through project " +
-                    "hierarchy). This means that $type closure was applied too early - before pom plugin created " +
-                    'convention in sub module. ' +
-                    "Most likely, this is because java plugin applied only in submodules section and this $type " +
-                    "closure is in allprojects section. To fix this, wrap $type section with afterEvaluate " +
-                    'to delay configuration until subproject convention would be created: \n' +
-                    '\tafterEvaluate { \n' +
-                    "\t\t$type { ... }\n" +
-                    '\t}')
-        }
+    void withPom(Closure config) {
+        this.rawConfigs.add(config)
     }
 
     /**
-     * Closures are hierarchical. Need to only find reference to the top closure, representing project.
+     * Raw xml modification action. This is the same as "pom.withXml" applied in publication, but this block would
+     * be applied in all registered publications.
      *
-     * @param config configuration closure (from build)
-     * @return project path
+     * @param modifier manual pom modification closure
      */
-    @CompileStatic(TypeCheckingMode.SKIP)
-    @SuppressWarnings('Instanceof')
-    private String findDeclarationProjectPath(Closure config) {
-        Object own = config.owner
-        while (own instanceof Closure && !own.hasProperty('path')) {
-            own = ((Closure) own).owner
-        }
-        // assuming this to be a project name
-        return own?.path
+    void withPomXml(Action<? super XmlProvider> modifier) {
+        this.xmlModifiers.add(modifier)
     }
 }