+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/DependencyManagementPluginKitTest.groovy
@@ -148,8 +148,8 @@ class DependencyManagementPluginKitTest extends AbstractKitTest {
             maven.pom {
                 developers {
                     developer {
-                        id 'jdoe'
-                        name 'John Doe'
+                        id = 'jdoe'
+                        name = 'John Doe'
                     }
                 }
             }