+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/NoDependenciesBlockKitTest.groovy
@@ -31,9 +31,9 @@ class NoDependenciesBlockKitTest extends AbstractKitTest {
             maven.pom {
                 developers {
                     developer {
-                        id "dev"
-                        name "Dev Dev"
-                        email "dev@gmail.com"
+                        id = "dev"
+                        name = "Dev Dev"
+                        email = "dev@gmail.com"
                     }
                 }
             }
@@ -96,9 +96,9 @@ class NoDependenciesBlockKitTest extends AbstractKitTest {
             maven.pom {
                 developers {
                     developer {
-                        id "dev"
-                        name "Dev Dev"
-                        email "dev@gmail.com"
+                        id = "dev"
+                        name = "Dev Dev"
+                        email = "dev@gmail.com"
                     }
                 }
             }