+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/WarCompatibilityTest.groovy
@@ -40,9 +40,9 @@ class WarCompatibilityTest extends AbstractKitTest {
             maven.pom {
                 developers {
                     developer {
-                        id "dev"
-                        name "Dev Dev"
-                        email "dev@gmail.com"
+                        id = "dev"
+                        name = "Dev Dev"
+                        email = "dev@gmail.com"
                     }
                 }
             }