+++ CHANGELOG.md
@@ -2,12 +2,17 @@
     - remove compile and runtime configurations support (were deprecated, now removed in gradle) 
 * (BREAKING) Conventions merged with extension (due to conventions deprecation):
     - pomGeneration extension renamed to maven
-    - pom and withPomXml conventions moved into extension: maven.pom and maven.withPomXml accordingly 
+    - old pom convention moved to maven.withPom (exactly the same as before)
+    - old withPomXml convention moved to maven withPomXml, but now its now an Action<XmlProvider>
+      (same as publication.pom.withXml) and so asNode() must now be called manually (instead of it)
+- Add static pom configuration: maven.pom. This is exactly the same as publication.pom (compatible with kotlin)
     
 Migration: 
 - Rename pomGeneration{} into maven{}
-- Rename pom{} and withPomXml{} into maven.pom{} and maven.withPomXml{} 
-  (or move then inside maven block, if pomGeneration extension was already used)
+- Rename pom{} into maven.withPom{} (exactly the same configuration) 
+    - Consider moving to type safe maven.pom - in this case all properties MUST contain '='
+      (this would be EXACTLY the same as pom configuration in publication)
+- Rename withPomXml{} into maven.withPomXml{} and replace 'it' with 'asNode()'
 
 ### 2.2.2 (2022-11-09)
 * Fix fail when repositories declared in settings file only (#35)